More Infinite potion effects

+++ Legion of Sins/src/main/java/me/waternode/server/legion/of/sins/mechanics/events/AcidRain.java
@@ -20,8 +20,8 @@ public class AcidRain extends CatastrophicEvent {
 	public void OnPlayerNear(Player p) {
 		World w = p.getWorld();
 		Location pLoc = p.getLocation();
-		Integer pX = (Integer) pLoc.getBlockX();
-		Integer pZ = (Integer) pLoc.getBlockZ();
+		Integer pX = pLoc.getBlockX();
+		Integer pZ = pLoc.getBlockZ();
 		if (w.getHighestBlockYAt(pX, pZ) - 5 <= pLoc.getY())
 			p.addPotionEffect(new PotionEffect(PotionEffectType.POISON, 120, 1));
 		p.getWorld().playSound(p.getLocation(), Sound.AMBIENCE_RAIN, 0.5f, 1f);
